requirements.txt
python-dotenv
Flask~=1.1.2
flask-graphql
flask-migrate
flask-sqlalchemy
graphene~=2.1.8
sqlalchemy~=1.3.15
graphene-sqlalchemy
pytest
factory_boy
python-dateutil~=2.8.1
gunicorn==20.0.4
psycopg2
sqlalchemy-utils